refactor(chat): extract stream_generator into chat_streaming.py
Move the 521-line SSE stream_generator (+ MAX_STREAM_DURATION deadline) out of
chat.py into a focused module, re-imported for src.routes.chat.stream_generator
patch-compat. chat.py 2716 -> ~2180 (-20% this step; -33% across 0d). Drops
imports orphaned by the move (StreamNormalizer, track_time_to_first_chunk,
create_done_sse, create_error_sse_chunk). Verbatim move, zero behavior change —
verified against the FULL local test suite (zero regressions vs baseline).

Also fixes test_streaming_background_task_only_after_done_event to read the new
module and assert the [DONE]->background-task order correctly (it previously
passed coincidentally by matching a handler create_task further down chat.py).
